Move Request.asCurl implementation in the CurlConverter
